refactor(lib): split lib.rs into config, ipc and service_ctl modules
Pure code motion: config structs, defaults and validation move to
config/types.rs, the IPC protocol to ipc.rs, and the systemd restart
helper to service_ctl.rs. lib.rs keeps WhisrsError and re-exports
everything at the crate root so call sites are unchanged.
