+++ src/config/mod.rs
@@ -2,3 +2,6 @@
 
 pub mod edit;
 pub mod setup;
+pub mod types;
+
+pub use types::*;
